+++ src/component/statistics/statistics.stories.tsx
@@ -1,12 +1,9 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import { Bookshelf } from '../../bookshelf/bookshelf'
-import { BookMetadataFactory } from '../../bookshelf/metadata/book-metadata-factory'
 import { FakeNote } from '../../support/fake-note'
 import { StaticMetadata } from '../../bookshelf/metadata/metadata'
-import { bookNotePatterns } from '../../bookshelf/reading-journey/pattern/book-note/book-note-pattern'
-import { dailyNotePatterns } from '../../bookshelf/reading-journey/pattern/daily-note/daily-note-pattern'
 import { Statistics } from './statistics'
 import { fn } from '@storybook/test'
+import { BookshelfFactory } from '../../bookshelf/bookshelf-factory'
 
 const meta = {
     title: 'Statistics',
@@ -15,45 +12,46 @@ const meta = {
 export default meta
 type Story = StoryObj<typeof Statistics>
 
-const bookshelf = new Bookshelf(
-    'Books',
-    {
-        heading: 'Reading Journey',
-        format: 'YYYY-MM-DD',
-        folder: '',
-    },
-    {
-        heading: 'Reading',
-    },
-    new BookMetadataFactory(
-        {
+const bookshelf = BookshelfFactory.fromConfiguration({
+    settings: {
+        booksFolder: 'Books',
+        bookProperties: {
             cover: 'cover',
             author: 'author',
             published: 'published',
             tags: 'tags',
             rating: 'rating',
         },
-        (link) => link,
-    ),
-    bookNotePatterns(
-        {
-            started: '{date}: Started reading',
-            abandoned: '{date}: Abandoned book',
-            finished: '{date}: Finished reading',
-            absoluteProgress: '{date}: {startPage}-{endPage}',
-            relativeProgress: '{date}: {endPage}',
+        bookNote: {
+            heading: 'Reading Journey',
+            dateFormat: 'yyyy-MM-dd',
+            patterns: {
+                started: '{date}: Started reading',
+                abandoned: '{date}: Abandoned book',
+                finished: '{date}: Finished reading',
+                absoluteProgress: '{date}: {startPage}-{endPage}',
+                relativeProgress: '{date}: {endPage}',
+            },
         },
-        'yyyy-MM-dd',
-    ).patterns,
-    dailyNotePatterns({
-        started: 'Started reading {book}',
-        abandoned: 'Abandoned {book}',
-        finished: 'Finished reading {book}',
-        absoluteProgress: 'Read {book}: {startPage}-{endPage}',
-        relativeProgress: 'Read {book}: {endPage}',
-    }).patterns,
-    (identifier) => identifier,
-)
+        dailyNote: {
+            heading: 'Reading',
+            patterns: {
+                started: 'Started reading {book}',
+                abandoned: 'Abandoned {book}',
+                finished: 'Finished reading {book}',
+                absoluteProgress: 'Read {book}: {startPage}-{endPage}',
+                relativeProgress: 'Read {book}: {endPage}',
+            },
+        },
+    },
+    dailyNotesSettings: {
+        enabled: true,
+        format: 'YYYY-MM-DD',
+        folder: '',
+    },
+    bookIdentifier: (input) => input,
+    linkToUri: (link) => link,
+})
 
 bookshelf.process(
     new FakeNote(